Responds to requests to create a new challenge update
The application is designed to respond to GitHub webhook requests.
At this point, the `POST "/challenge-updates"` route handler
verifies that a request came from GitHub and responds accordingly.
In the future it will need to parse the payload and send a message
to Slack.

The Dev Bootcamp Sinatra skeleton was used as the starting point
for this application.