Responds appropriately depending on whether a pull request was merged
GitHub will send requests for opening and closing pull requests
whether the pull request was merged or not.  So, the app needs to
verify that the request was sent in response to a pull request
being merged.

Next, it needs to verify that the repository was a challenge and
not a phase guide or other resource.